possible proxy fix for dalicc request
